Removed duplicate root folder
